backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
Pillow==12.0.0
celery==5.3.4
redis==5.0.1
python-decouple==3.8
django-celery-beat==2.5.0
django-filter==23.5